chore: app management typescript (#323)
