+++ package.json
@@ -19,7 +19,7 @@
   "scripts": {
     "clean": "rimraf dist",
     "build": "npm run clean && npm run build:modules && npm run build:standalone && npm run build:types",
-    "build:modules": "BABEL_ENV=modules babel lib -d dist/es-modules/",
+    "build:modules": "BABEL_ENV=modules babel --extensions \".ts\" --extensions \".js\" lib -d dist/es-modules/",
     "build:standalone": "webpack && NODE_ENV=production webpack",
     "build:standalone:log": "NODE_ENV=production WEBPACK_MODE=log webpack --json --profile --progress > webpack-build-log.json && webpack-bundle-analyzer webpack-build-log.json",
     "build:types": "rimraf typings/generated && tsc --declaration --noEmit false --emitDeclarationOnly true --allowJs false",
